refactor(handleMovieRequest): request 핸들하는 함수 추상화
